requirements.txt
django
django_rest_framework
psycopg2
coverage==3.6